map/requirements.txt
Django==4.0.6
pycopg2
djangorestframework
djangorestframework-jwt
django-redis







